# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CareNear API",
        "description": "CareNear backend API — Plateforme de mise en relation patients / infirmiers à domicile.",
        "contact": {
            "email": "contact@carenear.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Local dev server"
        }
    ],
    "paths": {
...
        "/nurses/search": {
            "get": {
                "tags": [
                    "Nurse Search"
                ],
                "summary": "Search nearby nurses by geolocation",
                "description": "Returns up to 50 verified nurses within the given radius, sorted by distance. Includes average rating and review count.",
                "operationId": "65e5f2565ea8832c72b1f2a89886f07c",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude of search center",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double",
                            "example": 18.102
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitude of search center",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double",
                            "example": -15.958
                        }
                    },
                    {
                        "name": "radius_km",
                        "in": "query",
                        "description": "Search radius in km (default 10, max 200)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 200,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 80
                        }
                    },
                    {
                        "name": "verified",
                        "in": "query",
                        "description": "Filter by verification status (default true)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Nouakchott"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Tevragh Zeina"
                                                    },
                                                    "lat": {
                                                        "type": "number",
                                                        "format": "double",
                                                        "example": 18.102
                                                    },
                                                    "lng": {
                                                        "type": "number",
                                                        "format": "double",
                                                        "example": -15.958
                                                    },
                                                    "coverage_km": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "price_min": {
                                                        "type": "integer",
                                                        "example": 500
                                                    },
                                                    "price_max": {
                                                        "type": "integer",
                                                        "example": 2000
                                                    },
                                                    "verified": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "distance_km": {
                                                        "type": "number",
                                                        "format": "double",
                                                        "example": 2.45
                                                    },
                                                    "rating_avg": {
                                                        "type": "number",
                                                        "format": "double",
                                                        "example": 4.5
                                                    },
                                                    "reviews_count": {
                                                        "type": "integer",
                                                        "example": 12
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error (lat/lng required)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patient/profile/documents/{index}/download": {
            "get": {
                "tags": [
                    "Patient Profile"
                ],
                "summary": "Download a specific medical document",
                "description": "PATIENT downloads own document by index. ADMIN can specify ?user_id=XX. Returns binary file.",
                "operationId": "25f0fbdc34cafe73c6ffb59043f83131",
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "description": "0-based document index",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "(ADMIN only) Target patient user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File download (binary stream)",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Sanctum token as Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}